+++ packages/terra-form-input/tests/wdio/form-input-spec.js
@@ -71,6 +71,30 @@ describe('Form-Input', () => {
       });
     });
 
+    describe('Enabled Placeholder', () => {
+      beforeEach(() => {
+        browser.url('/#/raw/tests/terra-form-input/form-input/placeholder-form-input');
+      });
+
+      Terra.should.beAccessible({ viewports });
+      Terra.should.matchScreenshot({ viewports });
+      Terra.should.themeEachCustomProperty({
+        '--terra-form-input-placeholder-color': 'red',
+      });
+    });
+
+    describe('Enabled Value', () => {
+      beforeEach(() => {
+        browser.url('/#/raw/tests/terra-form-input/form-input/value-form-input');
+      });
+
+      Terra.should.beAccessible({ viewports });
+      Terra.should.matchScreenshot({ viewports });
+      Terra.should.themeEachCustomProperty({
+        '--terra-form-input-color': 'red',
+      });
+    });
+
     describe('Enabled Keyboard-Focused', () => {
       before(() => {
         browser.url('/#/raw/tests/terra-form-input/form-input/default-form-input');
@@ -110,6 +134,36 @@ describe('Form-Input', () => {
           '--terra-form-input-disabled-border-width': '20rem dashed red',
           '--terra-form-input-disabled-color': 'red',
           '--terra-form-input-disabled-opacity': '0.5',
+        },
+      });
+    });
+
+    describe('Disabled Placeholder', () => {
+      before(() => {
+        browser.url('/#/raw/tests/terra-form-input/form-input/disabled-placeholder-input');
+      });
+
+      Terra.should.beAccessible({ viewports });
+      Terra.should.matchScreenshot({ viewports });
+      Terra.should.themeCombinationOfCustomProperties({
+        testName: 'themed',
+        properties: {
+          '--terra-form-input-placeholder-disabled-color': 'red',
+        },
+      });
+    });
+
+    describe('Disabled Value', () => {
+      before(() => {
+        browser.url('/#/raw/tests/terra-form-input/form-input/disabled-value-input');
+      });
+
+      Terra.should.beAccessible({ viewports });
+      Terra.should.matchScreenshot({ viewports });
+      Terra.should.themeCombinationOfCustomProperties({
+        testName: 'themed',
+        properties: {
+          '--terra-form-input-disabled-color': 'red',
         },
       });
     });